requirements.txt
flask>=3.0
gunicorn>=21.0
pillow>=10.0
pillow-heif>=0.18
